requirements.txt
annotated-types==0.5.0
anyio==3.7.1
chardet==5.1.0
click==8.1.6
exceptiongroup==1.1.2
fastapi==0.100.0
h11==0.14.0
httptools==0.6.0
idna==3.4
imbox==0.9.8
pydantic==2.1.1
pydantic_core==2.4.0
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
